tests: improved usability of TestKeys
